+++ src/lib.rs
@@ -4,4 +4,5 @@ pub mod dtb;
 pub mod gdb;
 pub mod loader;
 pub mod memory;
+pub mod snapshot;
 pub mod vm;